requirements.txt
numpy==2.4.1
pandas==3.0.0
python-dateutil==2.9.0.post0
ruff==0.14.14
six==1.17.0
fpdf2==2.7.9 # Added for PDF generation
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3
python-dotenv==1.0.1
